Restore `io.grpc` package name to match version pushed to Maven Central
Revert "bump version to 1.0.0 (#9)"

This reverts commit 95395410790dcc9ae05a861f218eff7cc11d59d9.

Revert "change groupid to `io.grpc.tools` (#7)"

This reverts commit 915b3d1c5663766e4318106d66fb2069623e8c60.

Revert "Revert "Bump version to 1.0.0 (#4)" (#6)"

This reverts commit 5705da71af55b5d3637145cfd7a1105e72ce702c.
