+++ pom.xml
@@ -17,7 +17,7 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>io.grpc.tools</groupId>
+  <groupId>io.grpc</groupId>
   <artifactId>grpc-java-api-checker</artifactId>
   <version>1.0.0</version>
